Make sure that combined traversals work, split traversals for single and multi-series
Queries need to work for several cases:

1. When we combine series, sum
2. When we map series, absolute
3. When we reduce series, derivative
4. Combination of any

Since (1) is different from all the others, in the way it combines
series and has to do (apply mapv ...), we have to make a separate
function for it. FWIW, reducer and mapper _might_ be safely removed
from there.

To handle (2) and (3) in simpler way, `traverse-each!` was added,
that handles operations on singlular series.